students/K33392/Taiakin_Daniil/lr1/requirements.txt
uvicorn[standard]==0.29.0
fastapi==0.110.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
aiohttp==3.9.5
sqlalchemy==2.0.30